add guard

+++ chronos_backend_app/src/domain/game_session/controller/game_session_player.controller.ts
@@ -1,27 +1,22 @@
-import { Body, Controller, Delete, NotFoundException, Post, UseGuards } from '@nestjs/common';
+import { BadRequestException, Body, Controller, Delete, NotFoundException, UseGuards } from '@nestjs/common';
 
-import { XSSPipe } from '../../../core/toolkit/pipe/xss.pipe';
-
-import { AuthGuard } from '../../guard/auth.guard';
-import { GameSessionPlayer } from '../model/game_session_player.model';
+import { AuthGuard } from '../../toolkit/guard/auth.guard';
+import { PlayerSessionGuard } from '../../toolkit/guard/player_session.guard';
 import { GameSessionPlayerService } from '../service/game_session_player.service';
 import { GameSessionPlayerDto } from '../dto/game_session_player.dto';
 
 @Controller('game_session_player')
-@UseGuards(AuthGuard)
+@UseGuards(AuthGuard, PlayerSessionGuard)
 export class GameSessionPlayerController  {
 
     constructor(private readonly gameSessionPlayerService: GameSessionPlayerService) {}
 
-    @Post()
-    async addGameSessionPlayer(@Body(XSSPipe) data: GameSessionPlayerDto): Promise<GameSessionPlayer | Error> {
-        const createdGameSessionPlayer = await this.gameSessionPlayerService.addGameSessionPlayer(data);
-
-        return createdGameSessionPlayer;
-    }
-
     @Delete()
     async removeGameSessionPlayer(@Body() data: GameSessionPlayerDto): Promise<void | Error> {
+        if(!data.gameSessionId) {
+            throw new BadRequestException();
+        }
+
         const deleteGameSessionPlayer = await this.gameSessionPlayerService.removeGameSessionPlayer(data.playerId, data.gameSessionId);
 
         if(deleteGameSessionPlayer === null) {
